% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfnfldatatools.R
\name{get_rush_type}
\alias{get_rush_type}
\title{NFL Team Rushing Tendencies}
\usage{
get_rush_type(season, run_type, start_week, end_week, graph = TRUE)
}
\arguments{
\item{start_week}{int: indicates the starting week of the selected season}

\item{end_week}{int: indicates the ending week of the selected season}

\item{graph}{boolean: indicates if you wish to graph output or return df, default is true}

\item{season_s}{int: indicates the season of interest}

\item{run_Type}{str: indicates a column in nflfastR to group rushes on (ie. run_gap, quarter, run_location,etc)}
}
\value{
df: dataframe containing the team, ypa, run type, attempts and epa
}
\description{
A wrapper function that groups rushing plays by offenses and rushing types
}
\examples{
\dontrun{
get_rush_type(2020, run_gap, 1, 17, graph = TRUE)
}
}
